Rename tenant -> tenants in routes
