+++ server/app.js
@@ -9,7 +9,7 @@ const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
 const propertiesRouter = require('./routes/properties');
 const workersRouter = require('./routes/workers');
-const tenantsRouter = require('./routes/tenant')
+const tenantsRouter = require('./routes/tenants')
 
 const app = express();
 app.use(cors());
